add initial test suite for ESignatureSerializer
