+++ server/apps/trak/tests/serializers/test_manifest_serializers.py
@@ -1,25 +1,5 @@
-import json
-import logging
-
-from django.test import TestCase
-
-from apps.trak.models import Handler, Manifest, WasteLine
+from apps.trak.models import Manifest, WasteLine
 from apps.trak.serializers import ManifestSerializer
-
-
-class SerializerBaseTests(TestCase):
-    def __int__(self, test_json, serializer):
-        self.test_json = test_json
-        self.serializer = serializer
-
-    def setUp(self) -> None:
-        self.valid = self.serializer.is_valid()
-        if not self.valid:
-            logging.error(f"{self.serializer.errors}")
-            self.fail(
-                f"{self.__class__.__name__} failed to initiate valid data\n"
-                f"{json.dumps(self.serializer.errors)}"
-            )
 
 
 class TestManifestSerializer:
@@ -43,12 +23,3 @@ class TestManifestSerializer:
         saved_manifest = manifest_10003114elc_serializer.save()
         waste_line = WasteLine.objects.filter(manifest=saved_manifest).first()
         assert isinstance(waste_line, WasteLine)
-
-
-class TestHandlerSerializer:
-    def test_save(self, handler_serializer):
-        if handler_serializer.is_valid():
-            saved_site = handler_serializer.save()
-            assert isinstance(saved_site, Handler)
-        else:
-            assert False
