+++ server/apps/trak/tests/serializers/test_signature_serializer.py
@@ -1,0 +1,21 @@
+from apps.trak.models import Signer
+from apps.trak.serializers.signature_ser import ESignatureSerializer
+
+
+class TestESignatureSerializer:
+    """
+    Test suite for e-Manifest electronic signatures serialization to/from JSON
+    """
+
+    def test_e_signature_serializes(self, e_signature_json):
+        e_signature_serializer = ESignatureSerializer(data=e_signature_json)
+        assert e_signature_serializer.is_valid() is True
+
+    def test_serializer_saves_new_signer(self, manifest_gen, e_signature_serializer, e_signature_json):
+        e_signature_serializer.save(manifest_handler=manifest_gen)
+        assert Signer.objects.filter(first_name=e_signature_json["signer"]["firstName"]).exists()
+
+    def test_e_signature_serializer_saves_new_document(self):
+        # ToDo: implement EPA's humanReadableDocument that's part
+        #  of the electronicSignaturesInfo body
+        pass
