+++ server/requirements_dev.txt
@@ -1,7 +1,7 @@
 pre-commit==3.1.1
-pytest==7.2.1
+pytest==7.2.2
 pytest-django==4.5.2
 responses==0.22.0
-ruff==0.0.253
+ruff==0.0.254
 black==23.1.0
 -r requirements.txt
